adding renamed of folders

+++ src/client/client.service.ts
@@ -64,13 +64,15 @@ export class ClientService {
       const currentClient: ClientEntity = await this.findByID(client.ClientID);
       const oldPath: string = path.join(__dirname, '..', '..', './files', currentClient.URL);
 
-      client.URL = client.Name.replace(/ /g, '-');
-      const newPath: string = path.join(__dirname, '..', '..', './files', client.URL);
+      if(fs.existsSync(oldPath)){
+        client.URL = client.Name.replace(/ /g, '-');
+        const newPath: string = path.join(__dirname, '..', '..', './files', client.URL);
 
-      fs.rename(oldPath, newPath, (error) => {
-        throw new HttpException('Error al renombrar carpeta de archivos del client', 500)
-      });
-      
+        fs.rename(oldPath, newPath, (error) => {
+          console.log(error)
+          throw new HttpException('Error al renombrar carpeta de archivos del client', 500)
+        });
+      }
   
     return await this.clientRepository.update(
         { ClientID: client.ClientID },
